# shared/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ApplicationListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ApplicationData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/ApplicationListMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApplicationListResponse",
        "type": "object"
      },
      "ApplicationResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ApplicationData"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ApplicationResponse",
        "type": "object"
      },
...
      "CreateApplicationRequest": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "default": "saved",
            "maxLength": 32,
            "title": "Status",
            "type": "string"
          },
          "vacancy_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Vacancy Id",
            "type": "string"
          }
        },
        "required": [
          "vacancy_id"
        ],
        "title": "CreateApplicationRequest",
        "type": "object"
      },
...
      "EventListResponse": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/EventData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Meta",
            "type": "object"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "EventListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-VacancyPilot-Client",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "VacancyPilot Ops Companion",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/applications": {
      "get": {
        "operationId": "list_applications_api_v1_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Applications",
        "tags": [
          "applications",
          "followups"
        ]
      },
      "post": {
        "operationId": "create_application_api_v1_applications_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Application",
        "tags": [
          "applications",
          "followups"
        ]
      }
    },
    "/api/v1/applications/{application_id}": {
      "patch": {
        "operationId": "update_application_api_v1_applications__application_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-VacancyPilot-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Vacancypilot-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApplicationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplicationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Application",
        "tags": [
          "applications",
          "followups"
        ]
      }
    },
    "/api/v1/applications/{application_id}/events": {
      "get": {
        "operationId": "list_application_events_api_v1_applications__application_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "List Application Events",
        "tags": [
          "applications",
          "followups"
        ]
      },
      "post": {
        "operationId": "create_application_event_api_v1_applications__application_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-VacancyPilot-Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Vacancypilot-Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Create Application Event",
        "tags": [
          "applications",
          "followups"
        ]
      }
    },
...
    "/api/v1/followups/{followup_id}": {
      "patch": {
        "operationId": "update_followup_api_v1_followups__followup_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "followup_id",
            "required": true,
            "schema": {
              "title": "Followup Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFollowUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Update Followup",
        "tags": [
          "applications",
          "followups"
        ]
      }
    },
...
  }
}